+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * fixed the issue of broken external links
 * improved link style
 * support marked content rendering: `==content==` → ==content==
+* improved toc scroll behaviour
+* support markdown footnote
 ## v0.2.0
 * added latex use case to post: theme style preview
 * added header/footer option to template:default, in order to hide header/footer sections on certain pages